pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "b1sl-python"
version = "0.1.1"
description = "A framework-agnostic Python SDK for SAP Business One Service Layer and SAP HANA. (Unofficial)"
readme = "README.md"
requires-python = ">=3.11,<3.13"
license = { text = "MIT" }
keywords = ["sap", "sap-b1", "service-layer", "hana", "erp", "odata"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",

    "Typing :: Typed",
]
authors = [
    { name = "Eliceo Guzman" },
]

# ── Core: minimal footprint ─────────────────────────────────────────────── #
dependencies = [
    "pydantic>=2.0.0,<3",
    "httpx>=0.28.1,<1",
    "python-dotenv>=1.2.2",
]

[project.urls]
Homepage = "https://github.com/operator-ita/b1sl-python"
Repository = "https://github.com/operator-ita/b1sl-python"
"Bug Tracker" = "https://github.com/operator-ita/b1sl-python/issues"
Documentation = "https://github.com/operator-ita/b1sl-python/tree/main/docs"

[project.optional-dependencies]
hana = ["hdbcli>=2.20.0,<3"]
django = ["django>=4.2,<6"]
generator = [
    "beautifulsoup4>=4.14.3",
    "lxml>=6.0.2",
]
dev = [
    "pytest>=8.0,<9.0",
    "pytest-django",
    "pytest-asyncio>=1.3.0",
    "pytest-recording>=0.13.4",
    "python-dotenv>=1.2.2",
    "respx>=0.22.0",
    "responses>=0.26.0",
    "ruff",
    "mypy",

    "pytest-cov",
]

all = ["b1sl-python[hana,django,generator]"]

[tool.hatch.build.targets.wheel]
packages = ["src/b1sl"]
exclude = [
    "/scripts",
    "/metadata",
    "/examples",
    "/docs",
]

[tool.hatch.build.targets.sdist]
exclude = [
    "/scripts",
    "/metadata",
    "/examples",
    "/docs",
    "/.env",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
markers = [
    "integration: tests that require a real SAP B1 Service Layer connection",
    "vcr: tests that use pytest-recording to capture HTTP traffic",
]
addopts = "--tb=short --cov=src --cov-report=term-missing"

[tool.mypy]
files = ["src"]
strict = true
ignore_missing_imports = true

[tool.ruff]
line-length = 88
target-version = "py311"
extend-exclude = [
    "src/b1sl/b1sl/models/_generated",
    "src/b1sl/b1sl/resources/_generated",
    "scripts",
]

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501", "W291", "W293"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ruff.lint.extend-per-file-ignores]
"src/b1sl/b1sl/entities/__init__.py" = ["F401", "F403", "F405"]
"examples/**" = ["E402", "F841"]

[tool.pyright]
include = ["src", "scripts"]
extraPaths = [".", "scripts"]
